+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.2
+####Fixed
+* P.A. Bolzano and P.A. Trento are showed as Trentino-Alto Adige from now on map and bar graph in order to conform with 
+the source of data used by this project 
 # 1.0.1
 ####Fixed
 * Fix visualization bug in Contatti frame on mobile versions